add deep lin to bucket activity
