+++ app/src/main/AndroidManifest.xml
@@ -17,10 +17,17 @@
             </intent-filter>
         </activity>
         <activity android:name=".activity.FTUActivity" />
-        <activity android:name=".activity.BucketActivity" />
+        <activity android:name=".activity.BucketActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+                <data android:scheme="wally" android:host="bucket" android:pathPrefix="/detail" />
+            </intent-filter>
+        </activity>
         <activity android:name=".activity.AddBucketEntryActivity"/>
         <activity android:name=".activity.AddBucketActivity"/>
-        <activity android:name=".activity.BucketListActivity"/>"
+        <activity android:name=".activity.BucketListActivity"/>
     </application>
 
 </manifest>
